requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.9.0
cryptography==43.0.1
pydantic==2.8.2
httpx==0.27.2
pytest==8.3.2
pytest-cov==5.0.0
ruff==0.6.9
black==24.8.0
